refactoring: transaction manager PR #12203
