+++ pkg/models/transaction.go
@@ -1,0 +1,7 @@
+package models
+
+import "context"
+
+type TransactionManager interface {
+	InTransaction(ctx context.Context, fn func(ctx context.Context) error) error
+}
